Updated test files

+++ nachos/test/unlink_test2.c
@@ -9,13 +9,18 @@ int main() {
         exit(1);
     }
     
+    int close_valid = close(fd);
+    if (close_valid == -1) {
+        printf("failed to close helloWorld.bin\n");
+        exit(1);
+    }
+    
     int fd_unlink = unlink("helloWorld.bin");
     
     if(fd_unlink != 0) {
         printf("Unlink failed\n");
         exit(1);
     }
-    
 }
